refactor: convert Integer to Double

+++ src/main/java/com/sztu/covid19/entity/VirusResult.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class VirusResult {
 
     private String date;
-    private Integer totalCT;
+    private Double totalCT;
 }
